Optimize advection
Performance of advection has improved significantly. For advection of
vector fields, the back trace location was being recomputed for each
component of vec field, now it will only do it once
(however the interpolation is still repeated).

Previously large amount of time was wasted on garbage collection of
the vectors in interpolation functions, StaticArrays are now used to
avoid wasting time with managing heap.